Remove not used alert controller factories.
